Allow puppetdb to be configure for masterless conf
Currently the module only allows PuppetDB to be configured in a
agent/master setup configuration, by configuring the master section of
the puppet.conf. This commit allows one to use this module to configure
the main section of the puppet.conf so the module can configure nodes to
use PuppetDB in a masterless setup.

Doc about puppet.conf configure in masterless way available here
https://docs.puppetlabs.com/puppetdb/2.2/connect_puppet_apply.html#manage-puppetconf